Azure Functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests==2.31.0
numpy==1.24.3
tensorflow-cpu==2.13.0
Pillow==9.5.0
